+++ src/library/DIPS.Mobile.UI/Resources/Styles/Chip/ChipStyleResources.cs
@@ -1,0 +1,14 @@
+/*
+Do not edit directly,
+generated 8/30/2023 10:47:37 AM from DIPS.Mobile.DesignTokens
+*/
+
+        namespace DIPS.Mobile.UI.Resources.Styles.Chip;
+        internal static class ChipStyleResources
+        {
+            public static Dictionary<ChipStyle, Style> Styles { get; } = new()
+            {
+                [ChipStyle.Input] = InputStyle.Current,
+                [ChipStyle.EmptyInput] = EmptyInputStyle.Current
+            };
+        }