+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [17.3.0]
 - Fixed an issue where ItemPicker in bottom sheet had strange behavior when searching.
+- [ItemPicker] The style when a placeholder is present has now changed.
+- [ItemPicker] The default Placeholder is now `Choose`. The previous value was string empty.
+- [ItemPicker] Fixed an issue where ItemPicker in bottom sheet had strange behavior when searching.
+- [ItemPicker] If the consumer changes ItemsSource when people have selected an item that is not a part of the new ItemsSource, the SelecteItem should be set to null.
+
 
 ## [17.2.0]
 - Resources was updated from DIPS.Mobile.DesignTokens